patch configurations

+++ HLTrigger/Configuration/python/HLT_75e33/modules/hltHgcalSoARecHitsLayerClustersProducer_cfi.py
@@ -7,9 +7,9 @@ hltHgcalSoARecHitsLayerClustersProducer = cms.EDProducer("HGCalCLUEsteringLayerC
     ),
     hgcalRecHitsSoA = cms.InputTag("hltHgcalSoARecHitsProducer"),
     detector = cms.string('EE'),
-    deltac = cms.float(1.3),
-    kappa = cms.float(9),
-    outlierDeltaFactor = cms.float(2.0)
+    deltac = cms.double(1.3),
+    kappa = cms.double(9),
+    outlierDeltaFactor = cms.double(2.0)
 )
 
 hltHgcalSoARecHitsLayerClustersProducerSerialSync = makeSerialClone(hltHgcalSoARecHitsLayerClustersProducer,
