+++ HLTrigger/Configuration/python/HLT_75e33/modules/hltHgcalSoARecHitsProducerL1Seeded_cfi.py
@@ -7,7 +7,7 @@ hltHgcalSoARecHitsProducerL1Seeded = cms.EDProducer("HGCalSoARecHitsProducer@alp
     ),
     dEdXweights = HGCAL_reco_constants.dEdXweights,
     detector = cms.string('EE'),
-    ecut = cms.double(3),
+    ecut = cms.float(3),
     fcPerEle = HGCAL_reco_constants.fcPerEle,
     fcPerMip = HGCAL_reco_constants.fcPerMip,
     maxNumberOfThickIndices = HGCAL_reco_constants.maxNumberOfThickIndices,
